backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
firebase-admin
python-dotenv